+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.7.0] - Unreleased
+## [0.7.0] - 2023-08-08
 ### Added
 - `function`, `import` and `export` tags [#9].
 - `runStringSync` function to run Vento in sync mode.
@@ -73,7 +73,7 @@ First version
 [#5]: https://github.com/oscarotero/vento/issues/5
 [#9]: https://github.com/oscarotero/vento/issues/9
 
-[0.7.0]: https://github.com/oscarotero/vento/compare/v0.6.0...HEAD
+[0.7.0]: https://github.com/oscarotero/vento/compare/v0.6.0...v0.7.0
 [0.6.0]: https://github.com/oscarotero/vento/compare/v0.5.1...v0.6.0
 [0.5.1]: https://github.com/oscarotero/vento/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/oscarotero/vento/compare/v0.4.0...v0.5.0